+++ tests/useAsyncFn.test.tsx
@@ -8,7 +8,7 @@
 import { act, renderHook } from '@testing-library/react-hooks';
 import useAsyncFn, { AsyncState } from '../src/useAsyncFn';
 
-type AdderFn = (a: number, b: number) => Promise<number>;
+type AdderFn = (a?: number, b?: number) => Promise<number>;
 
 describe('useAsyncFn', () => {
   it('should be defined', () => {
@@ -17,8 +17,8 @@ describe('useAsyncFn', () => {
 
   describe('the callback can be awaited and return the value', () => {
     let hook;
-    const adder = async (a: number, b: number): Promise<number> => {
-      return a + b;
+    const adder: AdderFn = async (a?: number, b?: number): Promise<number> => {
+      return (a || 0) + (b || 0);
     };
 
     beforeEach(() => {
@@ -50,9 +50,9 @@ describe('useAsyncFn', () => {
   describe('args can be passed to the function', () => {
     let hook;
     let callCount = 0;
-    const adder = async (a: number, b: number): Promise<number> => {
+    const adder = async (a?: number, b?: number): Promise<number> => {
       callCount++;
-      return a + b;
+      return (a || 0) + (b || 0);
     };
 
     beforeEach(() => {
@@ -93,5 +93,37 @@ describe('useAsyncFn', () => {
         expect(state.value).toEqual(9);
       });
     });
+  });
+
+  it('should only consider last call and discard previous ones', async () => {
+    const queuedPromises: { id: number; resolve: () => void }[] = [];
+    const delayedFunction1 = () => {
+      return new Promise<number>(resolve => queuedPromises.push({ id: 1, resolve: () => resolve(1) }));
+    };
+    const delayedFunction2 = () => {
+      return new Promise<number>(resolve => queuedPromises.push({ id: 2, resolve: () => resolve(2) }));
+    };
+
+    const hook = renderHook<{ fn: () => Promise<number> }, [AsyncState<number>, () => Promise<number>]>(
+      ({ fn }) => useAsyncFn(fn, [fn]),
+      {
+        initialProps: { fn: delayedFunction1 },
+      }
+    );
+    act(() => {
+      hook.result.current[1](); // invoke 1st callback
+    });
+
+    hook.rerender({ fn: delayedFunction2 });
+    act(() => {
+      hook.result.current[1](); // invoke 2nd callback
+    });
+
+    act(() => {
+      queuedPromises[1].resolve();
+      queuedPromises[0].resolve();
+    });
+    await hook.waitForNextUpdate();
+    expect(hook.result.current[0]).toEqual({ loading: false, value: 2 });
   });
 });